Move kuhn_munkers algorithm to common/utils
