+++ demos/pedestrian_tracker_demo/cpp/src/tracker.cpp
@@ -15,7 +15,7 @@
 #include "core.hpp"
 #include "tracker.hpp"
 #include "utils.hpp"
-#include "kuhn_munkres.hpp"
+#include "utils/kuhn_munkres.hpp"
 
 namespace {
 cv::Point Center(const cv::Rect& rect) {
